minecraft_client: Started basic login support
This patch adds the login start and login success packets, as well as
all the code for handling them as part of the login procedure.

See:
[wiki.vg's Protocol FAQ](https://wiki.vg/Protocol_FAQ#What.27s_the_normal_login_sequence_for_a_client.3F)
for more informations on the login procedure.

This also adds basic support for handling uuids, though it should be
noted it uses the 64bits-specific type `__int128`, so should probably be
rewritten better. I also didn't test this at all, so it might break.