minecraft_client: Send proper dimension and dimension_codec in join_game
This sends real NBT data for those fields, soon login will work
(hopefully!).

This data is loaded from disk by the minecraft_server, and passed to any
minecraft_client.

It seems like there is an issue in packet lenght calculation that is
preventing the game from noticing we've sent all the data...